refactor: create kafka helpers
This commit creates a new package `internal/kafka` that contains
helpers for sending and receiving messages from Kafka. The `send.go`
package is refactored to use these helpers.

This is in preparation for adding the `verify` implementation.
